refactor: Hash.h를 HashUtils.h로 이름 변경 및 관련 코드에 반영
