+++ EngineCore/Include/SimpleEngine/Utility/HashUtils.h
